requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
google-generativeai
httpx
pydantic
